moving FakeRedis to vumi.tests.utils
