+++ vumi/workers/failures/tests/test_workers.py
@@ -1,92 +1,15 @@
 import time
 import json
-import fnmatch
 from datetime import datetime, timedelta
 
 from twisted.trial import unittest
 from twisted.internet.defer import inlineCallbacks
 
-from vumi.tests.utils import get_stubbed_worker
+from vumi.tests.utils import get_stubbed_worker, FakeRedis
 from vumi.workers.failures.workers import FailureWorker
 
 
 RETRY_TIMESTAMPS_KEY = "failures:vas2nets#retry_timestamps"
-
-
-class FakeRedis(object):
-    def __init__(self):
-        self._data = {}
-
-    # Global operations
-
-    def exists(self, key):
-        return key in self._data
-
-    def keys(self, pattern='*'):
-        return fnmatch.filter(self._data.keys(), pattern)
-
-    def flushdb(self):
-        self._data = {}
-
-    # String operations
-
-    def get(self, key):
-        return self._data.get(key)
-
-    def set(self, key, value):
-        self._data[key] = value
-
-    # Hash operations
-
-    def hmset(self, key, mapping):
-        hval = self._data.setdefault(key, {})
-        hval.update(mapping)
-
-    def hgetall(self, key):
-        return self._data.get(key, {})
-
-    # Set operations
-
-    def sadd(self, key, value):
-        sval = self._data.setdefault(key, set())
-        sval.add(value)
-
-    def smembers(self, key):
-        return self._data.get(key, set())
-
-    def spop(self, key):
-        sval = self._data.get(key, set())
-        if not sval:
-            return None
-        return sval.pop()
-
-    def scard(self, key):
-        return len(self._data.get(key, set()))
-
-    # Sorted set operations
-
-    def zadd(self, key, **valscores):
-        zval = self._data.setdefault(key, [])
-        new_zval = [val for val in zval if val[1] not in valscores]
-        for value, score in valscores.items():
-            new_zval.append((score, value))
-        new_zval.sort()
-        self._data[key] = new_zval
-
-    def zrem(self, key, value):
-        zval = self._data.setdefault(key, [])
-        new_zval = [val for val in zval if val[1] != value]
-        self._data[key] = new_zval
-
-    def zcard(self, key):
-        return len(self._data.get(key, []))
-
-    def zrange(self, key, start, stop):
-        zval = self._data.get(key, [])
-        stop += 1  # redis start/stop are element indexes
-        if stop == 0:
-            stop = None
-        return [val[1] for val in zval[start:stop]]
 
 
 def mktimestamp(delta=0):
